feat: split it all up
