+++ src/tabs-motion/constants.js
@@ -1,0 +1,32 @@
+/*
+Copyright (c) 2018-2020 Uber Technologies, Inc.
+
+This source code is licensed under the MIT license found in the
+LICENSE file in the root directory of this source tree.
+*/
+
+// @flow
+
+export const ORIENTATION = {
+  vertical: 'vertical',
+  horizontal: 'horizontal',
+};
+
+export const FILL = {
+  fixed: 'fixed',
+  intrinsic: 'intrinsic',
+};
+
+export const KEYBOARD_ACTIVATION = {
+  automatic: 'automatic',
+  manual: 'manual',
+};
+
+export const KEYBOARD_ACTION = {
+  next: 'next',
+  previous: 'previous',
+};
+
+export const STATE_CHANGE_TYPE = {
+  change: 'change',
+};
